added more styling to notes section

+++ actnaturalapp/views/animals/animal_details.py
@@ -2,16 +2,18 @@ from django.urls import reverse
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from actnaturalapp.models import Animal, Team, Species, Employee, AnimalNote
+from django.contrib.auth.models import User
 
 
 @login_required
-def animal_details(request, animal_id):
+def animal_details(request, animal_id=None, note_id=None):
 
     animal = Animal.objects.get(pk=animal_id)
     team = Team.objects.get(pk=animal.team_id)
     species = Species.objects.get(pk=animal.species_id)
     employee = Employee.objects.get(pk=request.user.employee.id)
     notes = AnimalNote.objects.filter(animal_id=animal_id)
+    users = User.objects.all()
 
     if request.method == 'GET':
 
@@ -21,7 +23,8 @@ def animal_details(request, animal_id):
             'team': team,
             'species': species,
             'employee': employee,
-            'notes': notes
+            'notes': notes,
+            'users': users
         }
 
         return render(request, template, context)
@@ -51,11 +54,25 @@ def animal_details(request, animal_id):
         elif (
             "actual_method" in form_data
             and form_data["actual_method"] == "DELETE"
-        ):
-            
-            animal.delete()
+        ):  
 
-            return redirect(reverse('actnaturalapp:animals'))
+            if (
+                "age" in form_data
+            ):
+            
+                animal.delete()
+
+                return redirect(reverse('actnaturalapp:animals'))
+
+            elif (
+                "note" in form_data
+            ):
+
+                note = AnimalNote.object.get(pk=note_id)
+
+                note.delete()
+
+                return redirect(reverse('actnaturalapp:animal', args=[animal.id]))
 
         else:
